Added tiled image layer resolution switching.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@1082 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWindIOS/WorldWind/WorldWind/Util/WWLevelSet.h
@@ -90,7 +90,7 @@
 *
 * @return YES if the level number indicated the highest resolution level in the level set, otherwise NO.
 */
-- (BOOL) isLastLevel:(NSUInteger)levelNumber;
+- (BOOL) isLastLevel:(int)levelNumber;
 
 /// @name Initializing Level Sets
